move the kernel params tooling out to it's own file to be used elsewhere
